+++ src/SmartHopper.Core.Grasshopper/AITools/gh_smart_connect.cs
@@ -30,7 +30,6 @@ using Newtonsoft.Json.Linq;
 using Rhino;
 using SmartHopper.Core.Grasshopper.Utils.Canvas;
 using SmartHopper.Infrastructure.AICall.Tools;
-using SmartHopper.Infrastructure.AIModels;
 using SmartHopper.Infrastructure.AITools;
 using SmartHopper.ProviderSdk.AICall.Core.Base;
 using SmartHopper.ProviderSdk.AICall.Core.Interactions;