+++ src/SmartHopper.Infrastructure/AICall/Sessions/SpecialTurns/SpecialTurnConfig.cs
@@ -19,7 +19,6 @@
 namespace SmartHopper.Infrastructure.AICall.Sessions.SpecialTurns
 {
     using System.Collections.Generic;
-    using SmartHopper.Infrastructure.AIModels;
     using SmartHopper.ProviderSdk.AICall.Core.Interactions;
     using SmartHopper.ProviderSdk.AIModels;
     /// <summary>